+++ server_setup/__init__.py
@@ -1,9 +1,10 @@
 from server_setup.certificate import certificate
 from server_setup.nginx_site import nginx_site
 from server_setup.observability import observability
-from server_setup.postgres_cluster import postgres_backups, postgres_logging
+from server_setup.postgres_cluster import postgres_backups, postgres_config
 from server_setup.postgres_db import postgres_db
 from server_setup.system import firewall, nginx, packages, services, ssh, swap
+from server_setup.upgrade import upgrade
 
 __all__ = [
     "certificate",
@@ -13,9 +14,10 @@ __all__ = [
     "observability",
     "packages",
     "postgres_backups",
+    "postgres_config",
     "postgres_db",
-    "postgres_logging",
     "services",
     "ssh",
     "swap",
+    "upgrade",
 ]